Add checking for database emptiness on startup, and base for paging through people.
Now Asyncronously check emptiness of database in OnCreate of LoadActivity. Only allows user to view dossiers for people if the database is non-empty. Also add a fragment to display the information on each person in the database, and add the PeopleViewerActivity which will allow paging through these fragments in the next commit.