Fix bug where deleting while scrolling between people would crash the app.
I now handle deletion differently. When I start, I close the cursor on my adapter and set it to null. Whenever the adapters cursor is null, and either the deletion, or loading task is in progress, the adapter will display a single page of a "default loading person". Before, what I was doing was setting the ViewPager's adapter as null while I was updating the cursor, which would result in a crash if the User forced the view to update before I'd replaced it with a valid Adapter (as in the case of scrolling between pages).